Move DAGCell implementation into the dagmc files.
